+++ src/components/create-course-illustration.js
